feat: Add getPatientById function to usePatient hook
This commit adds a new function `getPatientById` to the `usePatient` hook. This function allows fetching a specific patient by their ID from the server. It makes a GET request to the API endpoint `/patients/{clinicId}/{patientId}` with the necessary authorization headers. If the request is successful, the patient data is returned. If there is an error, it is logged and rethrown.